requirements.txt
flask>=3.0.0
pydub>=0.25.1
requests>=2.31.0
python-dotenv>=1.0.0
shazamio>=0.4.0
